requirements/develop.txt
amqp==5.3.1
asgiref==3.8.1
autoflake==2.2.0
billiard==4.2.1
black==23.7.0
blinker==1.9.0
celery==5.5.3
certifi==2025.6.15
cffi==1.17.1
cfgv==3.4.0
channels==4.2.2
channels_redis==4.2.1
charset-normalizer==3.4.2
classify-imports==4.2.0
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cryptography==41.0.7
distlib==0.3.9
dj-database-url==2.2.0
Django==5.2.3
django-cors-headers==4.7.0
django-filter==25.1
django-jazzmin==3.0.1
django-modeltranslation==0.19.15
django-recaptcha==3.0.0
django-redis==6.0.0
django-rosetta==0.10.2
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.10
Faker==37.4.0
filelock==3.18.0
flake8==7.3.0
flower==2.0.1
humanize==4.12.3
identify==2.6.12
idna==3.10
inflection==0.5.1
isort==5.12.0
kombu==5.5.4
mccabe==0.7.0
msgpack==1.1.1
mypy_extensions==1.1.0
nodeenv==1.9.1
nplusone==1.0.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.3.8
polib==1.2.0
pre_commit==4.2.0
prometheus_client==0.22.1
prompt_toolkit==3.0.51
psycopg2-binary==2.9.10
pycodestyle==2.14.0
pycparser==2.22
pyflakes==3.4.0
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-environ==0.4.54
pytz==2025.2
pyupgrade==3.10.1
PyYAML==6.0.2
redis==5.0.1
reorder-python-imports==3.10.0
requests==2.32.4
six==1.17.0
sqlparse==0.5.3
tokenize_rt==6.2.0
tornado==6.5.1
typing_extensions==4.14.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
vine==5.1.0
virtualenv==20.31.2
wcwidth==0.2.13
yesqa==1.5.0
